+++ src/database/repositories/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const UserRepository = require('./user');
+const RecipeRepository = require('./recipe');
+const CategoryRepository = require('./category');
+
+module.exports = {
+  UserRepository,
+  RecipeRepository,
+  CategoryRepository,
+};
